Added active speaker hook
